Simple quality of life changes (speed and integer coordinates.

+++ src/game/util/Vector.java
@@ -55,7 +55,7 @@ public class Vector {
     
     @Override
     public String toString() {
-        return "[" + x + ", " + y + ", " + z + "]";
+        return "[" + (int)x + ", " + (int)y + ", " + (int)z + "]";
     }
 
     public float getMagnitude() {
